% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_datasets.R
\name{available_datasets}
\alias{available_datasets}
\title{Title}
\usage{
available_datasets(
  data_dir,
  species,
  datasets = "all",
  return_distribution = FALSE
)
}
\arguments{
\item{return_distribution}{}
}
\description{
Title
}
