% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_flags.R
\name{count_flags}
\alias{count_flags}
\title{Count the number of flags for each record}
\usage{
count_flags(
  occ = NULL,
  flagged_dir = NULL,
  output_format = ".gz",
  flags = "all",
  additional_flags = NULL
)
}
\arguments{
\item{occ}{(data.frame or data.table) a dataset containing occurrence records
that has been processed by one or more flagging functions. See \emph{Details} for
available flag types.}

\item{flagged_dir}{(character) optional path to a directory containing files
with flagged records saved using the \code{remove_flagged()} function. Default is
\code{NULL}.}

\item{output_format}{(character) output format used to read the removed records.
Options are \code{".csv"} or \code{".gz"}. Only used when \code{flagged_dir} is not \code{NULL}.
Default is \code{".gz"}.}

\item{flags}{(character) the flags to be summarized. Use \code{"all"} to display
all available flags. See \emph{Details} for all options. Default is \code{"all"}.}

\item{additional_flags}{(character) an optional named character vector with
the names of additional logical columns to be used as flags. Default is \code{NULL}.}
}
\value{
The original data frame with an additional column summarizing the total
number of flags triggered by each record.
}
\description{
This function returns the original data frame with an additional column
summarizing the total number of flags triggered by each record.
}
\details{
This function expects an occurrence dataset that has already been processed
by one or more flagging routines from \strong{RuHere} or related packages such as
\strong{CoordinateCleaner}. Any logical column in \code{occ} can be used as a flag.

The following built-in flag names are recognized:

\emph{From RuHere}:
\code{correct_country}, \code{correct_state}, \code{cultivated}, \code{florabr}, \code{faunabr},
\code{wcvp}, \code{iucn}, \code{bien}, \code{duplicated}, \code{thin_geo}, \code{thin_env}, \code{consensus}

\emph{From CoordinateCleaner} :
\code{.val}, \code{.equ}, \code{.zer}, \code{.cap}, \code{.cen}, \code{.sea}, \code{.urb}, \code{.otl}, \code{.gbf},
\code{.inst}, \code{.aohi}

Users may also supply additional logical columns using
\code{additional_flags}.
}
\examples{
# Load example data
data("occ_flagged", package = "RuHere")

# Count flags
sum_flags <- count_flags(occ = occ_flagged)

# Check the distribution of flags per record
table(sum_flags$total_flags)

# Plot histogram
hist(sum_flags$total_flags,
     main = "Distribution of Flags",
     xlab = "Number of Flags",
     col = "steelblue",
     breaks = seq(-0.5, max(sum_flags$total_flags) + 0.5, by = 1))
}
