% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase_here.R
\name{fishbase_here}
\alias{fishbase_here}
\title{fishbase_here}
\usage{
fishbase_here(data_dir, data_version = "latest", verbose = TRUE)
}
\arguments{
\item{data_dir}{(character) the path to the local directory where a new
fishbase subdirectory will be created to store all downloaded data files. \strong{Required.}}

\item{data_version}{(character) the specific version of FishBase data to query.
Default is \code{"latest"}.}

\item{verbose}{(logical) if \code{TRUE}, messages indicating the download
progress. Default is \code{TRUE}.}
}
\value{
A message indicating that the data were successfully saved in the directory
specified by \code{data_dir}.
}
\description{
Downloads taxonomic and distribution tables from the \strong{FishBase}
database using the \code{rfishbase} package and saves them locally in a
standardized directory structure.
}
\details{
All files are saved compressed as \code{.gz} files within a new \code{fishbase}
subdirectory inside \code{data_dir}.
}
\examples{
\dontrun{
# Define a directory to save the data
data_dir <- tempdir() # Here, a temporary directory

# Download the latest version of the FishBase database
fishbase_here(data_dir = data_dir)
}

}
