% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flag_colors}
\alias{flag_colors}
\title{Color palette for flagged records}
\format{
A named character vector where:

\describe{
\item{names}{Flag labels corresponding to categories generated by the
various \verb{flag_*} and checking functions.}

\item{values}{Hex color codes or standard R color names used for plotting.}
}
}
\usage{
flag_colors
}
\description{
\code{flag_colors} is a named character vector defining the default colors used to
plot occurrence records flagged with \code{mapview_here()}.
}
\examples{
data(flag_colors)

# View all flag categories and their colors
flag_colors

}
\seealso{
\code{mapview_here}
}
\keyword{datasets}
