% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flag_names}
\alias{flag_names}
\title{Flag name dictionary}
\format{
A named character vector of length 25.
The names correspond to the original flag codes (e.g., \code{"correct_country"},
\code{"duplicated_flag"}, \code{".cen"}, \code{"consensus_flag"}), and the values are the
cleaned, human-readable labels (e.g., \code{"Wrong country"}, \code{"Duplicated"},
\code{"Country/Province centroid"}, \code{"consensus"}).
}
\usage{
flag_names
}
\description{
A named character vector used to convert internal flag column names
(produced by the package's flagging functions) into human-readable labels.
}
\details{
This object is used internally by functions such as \code{mapview_here()} and
\code{remove_flagged()}to display more intuitive flag names to users.
}
\keyword{datasets}
