% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datapaper.R
\name{get_datapaper}
\alias{get_datapaper}
\title{Download and merge biodiversity data papers}
\usage{
get_datapaper(
  datapaper,
  species = NULL,
  save = FALSE,
  dir = NULL,
  file.format = "gz",
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{datapaper}{(character) vector with the names of the data papers to
download. Valid options are: \code{"atlantic"}, \code{"brazil road-kill"}, \code{"dryflor"},
\code{"neotropical xenarthrans"}, and \code{"neotroptree"}.}

\item{species}{(character) optional species name or vector of species names
to filter the dataset. Default is \code{NULL} (returns all species).}

\item{save}{(logical) whether to save the merged output file to disk. If
\code{FALSE}, files are downloaded to a temporary session directory. Default is
\code{FALSE}.}

\item{dir}{(character) directory path where the output file will be saved if
\code{save = TRUE}. Default is \code{NULL}.}

\item{file.format}{(character) file format to save the merged table. Valid
options are \code{"gz"} (compressed CSV), \code{"csv"}, or \code{"rds"}. Default is \code{"gz"}}

\item{overwrite}{(logical) whether to overwrite existing downloaded files in
the destination directory. Default is \code{FALSE}}

\item{verbose}{(logical) whether to display messages during function
execution. Set to TRUE to enable display, or FALSE to run silently. Default
is TRUE.}
}
\value{
A \code{data.table} containing the consolidated records from the selected
data papers.
}
\description{
Downloads, reads, and merges published biodiversity data papers (see details
for available options).
The function allows downloading single or multiple datasets simultaneously,
optionally filtering records by species name and saving the consolidated
table to disk in different file formats.
}
\details{
Available datasets that can be requested in \code{datapaper}:
\itemize{
\item \code{"atlantic"}: Atlantic data papers compilations.
\item \code{"brazil road-kill"}: Brazil road-kill dataset.
\item \code{"dryflor"}: Latin American and Caribbean Seasonally Dry
Tropical Forests (DryFlor) dataset.
\item \code{"neotropical xenarthrans"}: Neotropical Xenarthrans occurrence
dataset.
\item \code{"neotroptree"}: NeoTropTree tree species checklist database.
}
}
\note{
When using any dataset retrieved by this function in publications or research
reports, please cite the original authors. Use the function
\code{cite_datapaper()} to obtain the complete bibliographic references.
}
\examples{
 \dontrun{
# Download Atlantic dataset
res <- get_datapaper(datapaper = "atlantic")

# Download multiple datasets and filter by a specific species
res_sp <- get_datapaper(
  datapaper = c("atlantic", "dryflor"),
  species = "Araucaria angustifolia")
}
}
