% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occ_idig}
\alias{occ_idig}
\title{Occurrence records of azure jay from iDigBio}
\format{
A data frame containing georeferenced iDigBio occurrence records for
\emph{C. caeruleus} after all cleaning steps.
}
\usage{
occ_idig
}
\description{
A cleaned dataset of occurrence records for azure jay (\emph{Cyanocorax caeruleus})
retrieved from the iDigBio using \code{get_idigbio()}.

Records were cleaned using the package's internal duplicate-flagging workflow.
}
\examples{
# First rows
head(occ_idig)

# Number of cleaned records
nrow(occ_idig)

}
\seealso{
\code{get_idigbio()}, \code{flag_duplicates()}, \code{remove_flagged()}
}
\keyword{datasets}
