% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occ_splink}
\alias{occ_splink}
\title{Occurrence records of azure jay from SpeciesLink}
\format{
A data frame containing georeferenced SpeciesLink occurrence records for
\emph{C. caeruleus} after all cleaning steps.
}
\usage{
occ_splink
}
\description{
A cleaned dataset of occurrence records for azure jay (\emph{Cyanocorax caeruleus})
retrieved from the SpeciesLink using \code{get_specieslink()}.

Records were cleaned using the package's internal duplicate-flagging workflow.
}
\examples{
# First rows
head(occ_splink)

# Number of cleaned records
nrow(occ_splink)

}
\seealso{
\code{get_specieslink()}, \code{flag_duplicates()}, \code{remove_flagged()}
}
\keyword{datasets}
