% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_accent.R
\name{remove_accent}
\alias{remove_accent}
\title{Remove accents and special characters from strings}
\usage{
remove_accent(s)
}
\arguments{
\item{s}{(character) a character vector containing the strings to process.}
}
\value{
A vector string without accents or special characters
}
\description{
This function removes accents and replaces special characters from strings,
returning a plain-text version suitable for data cleaning or standardization.
}
\examples{
remove_accent(c("Colômbia", "São Paulo"))
}
