% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_flagged.R
\name{remove_flagged}
\alias{remove_flagged}
\title{Title}
\usage{
remove_flagged(
  occ,
  flags = "all",
  exceptions = NULL,
  column_exception = NULL,
  save_flagged = FALSE,
  output_dir = NULL,
  overwrite = FALSE,
  output_format = ".gz"
)
}
\arguments{
\item{output_format}{}
}
\description{
Title
}
