% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gbif_credentials.R
\name{set_gbif_credentials}
\alias{set_gbif_credentials}
\title{Store GBIF credentials}
\usage{
set_gbif_credentials(gbif_username, gbif_email, gbif_password,
                           verbose = TRUE)
}
\arguments{
\item{gbif_username}{(character) your GBIF username.}

\item{gbif_email}{(character) your GBIF email address.}

\item{gbif_password}{(character) your GBIF password.}

\item{verbose}{(logical) if \code{TRUE}, prints messages about the progress and
the number of species being checked. Default is \code{TRUE}.}
}
\value{
No return value.
}
\description{
This function sets GBIF credentials (username, email and password) as
environment variables for the current R session. These credentials are required
to retrieve occurrence records from GBIF.
}
\details{
To make these credentials permanent, you can manually add them to your
\code{.Renviron} file.
}
\examples{
\dontrun{
set_gbif_credentials(gbif_username = "my_username",
                     gbif_email = "my_email@example.com",
                     gbif_password = "my_password")
}
}
