% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_records.R
\name{thin_records}
\alias{thin_records}
\title{Title}
\usage{
thin_records(
  occ,
  species = "species",
  long = "decimalLongitude",
  lat = "decimalLatitude",
  d,
  prioritary_column = NULL,
  desc = TRUE
)
}
\arguments{
\item{desc}{}
}
\description{
Title
}
