% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_cultivated.R
\name{flag_cultivated}
\alias{flag_cultivated}
\title{Title}
\usage{
flag_cultivated(
  occ,
  columns = c("occurrenceRemarks", "habitat", "locality"),
  cultivated_terms = NULL,
  not_cultivated_terms = NULL
)
}
\arguments{
\item{not_cultivated_terms}{}
}
\description{
Title
}
