% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_duplicates.R
\name{flag_duplicates}
\alias{flag_duplicates}
\title{Title}
\usage{
flag_duplicates(
  occ,
  species = "species",
  long = "decimalLongitude",
  lat = "decimalLatitude",
  additional_groups = NULL,
  continuous_variable = NULL,
  decreasing = TRUE,
  categorical_variable = NULL,
  priority_categories = NULL,
  by_cell = FALSE,
  raster_variable = NULL
)
}
\arguments{
\item{raster_variable}{}
}
\description{
Title
}
