% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_fossil.R
\name{flag_fossil}
\alias{flag_fossil}
\title{Flag fossil}
\usage{
flag_fossil(
  occ,
  columns = c("basisOfRecord", "occurrenceRemarks"),
  fossil_terms = NULL
)
}
\arguments{
\item{occ}{(data.frame)}

\item{columns}{(character)}

\item{fossil_terms}{(character)}
}
\value{
A dataframe
}
\description{
Flag fossil
}
