% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_inaturalist.R
\name{flag_inaturalist}
\alias{flag_inaturalist}
\title{Title}
\usage{
flag_inaturalist(occ, columns = "datasetName", research_grade = FALSE)
}
\arguments{
\item{research_grade}{}
}
\description{
Title
}
