% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_iucn.R
\name{flag_iucn}
\alias{flag_iucn}
\title{Title}
\usage{
flag_iucn(
  data_dir,
  occ,
  species = "species",
  long = "decimalLongitude",
  lat = "decimalLatitude",
  origin = NULL,
  presence = NULL,
  buffer = 10,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
