% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_here.R
\name{ggmap_here}
\alias{ggmap_here}
\title{Static visualization of occurrence flags with ggplot2}
\usage{
ggmap_here(
  occ,
  long = "decimalLongitude",
  lat = "decimalLatitude",
  flags = "all",
  additional_flags = NULL,
  names_additional_flags = NULL,
  col_additional_flags = NULL,
  show_no_flagged = TRUE,
  col_points = NULL,
  size_points = 1,
  continent = NULL,
  continent_fill = "gray70",
  continent_linewidth = 0.3,
  continent_border = "white",
  ocean_fill = "aliceblue",
  extension = NULL,
  facet = FALSE,
  ...
)
}
\description{
Produces a static map using \strong{ggplot2}, highlighting flagged records.
}
