% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranfast.R
\name{moranfast}
\alias{moranfast}
\title{Title}
\usage{
moranfast(
  x,
  weight,
  na_rm = FALSE,
  scaled = FALSE,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{alternative}{}
}
\description{
Title
}
