% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_here.R
\name{plot_here}
\alias{plot_here}
\title{View occurrence flags interactively or with ggplot}
\usage{
plot_here(
  occ,
  flags = "all",
  show_no_flagged = TRUE,
  plot_with = "mapview",
  cex = 6,
  lwd = 2,
  col_points = NULL,
  label = NULL
)
}
\arguments{
\item{occ}{(data.frame or data.table) a dataset with occurrence records that was passed by one or more of the flag functions. See details.}

\item{flags}{(character) the flags used to color the records.}

\item{show_no_flagged}{(logical) whether to show the records with no flags. Default is TRUE.}

\item{plot_with}{(character)}

\item{cex}{(numeric) the size of the points representing occurrences. Default is 6.}

\item{lwd}{(numeric) line width of the points representing occurrences. Default is 2.}

\item{col_points}{(character) a named vector with the colors of the points representing each flag. If NULL, it will use the default colors provided in (\strong{cite object here})}

\item{label}{(character) the column name of \code{occ} with labels to be shown on mouseover. Default is NULL, meaning it will show the row.number of the record on mouseover.}
}
\value{
If \code{plot_with = "mapview"}, an interactive \code{mapview} object.
If \code{plot_with = "ggplot"}, a static \code{ggplot} object.
}
\description{
This function produces an interactive os static plot of occurrence records, highlighting the flags
}
