% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_invalid_coordinates.R
\name{remove_invalid_coordinates}
\alias{remove_invalid_coordinates}
\title{Identify and remove invalid coordinates}
\usage{
remove_invalid_coordinates(
  occ,
  long = "decimalLongitude",
  lat = "decimalLatitude",
  return_invalid = TRUE
)
}
\arguments{
\item{occ}{(data.frame or data.table) a dataset with occurrence records.}

\item{long}{(character) column name in \code{occ} with the longitude.}

\item{lat}{(character) column name in \code{occ} with the latitude.}

\item{return_invalid}{(logical) whether to return a list containing the valid
and invalid coordinates. Default is TRUE.}
}
\value{
If \code{return_invalid = FALSE}, returns the occurrence dataset containing only
valid coordinates.
If \code{return_invalid = TRUE} (default), returns a list with two elements:
\itemize{
\item \code{valid} – the dataset with valid coordinates.
\item \code{invalid} – the dataset with invalid coordinates removed.
}
}
\description{
This function identifies and removes invalid geographic coordinates,
including non-numeric values, NA or empty values, and coordinates outside the
valid range for Earth (latitude > 90 or < -90, and longitude > 180 or < -180).
}
\examples{
# Create fake data example
occ <- data.frame("species" = "spp",
                  "decimalLongitude" = c(10, -190, 20, 50, NA),
                  "decimalLatitude" = c(20, 20, 240, 50, NA))
# Split valid and invalid coordinates
occ_valid <- remove_invalid_coordinates(occ)
}
