% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_iucn_credentials.R
\name{set_iucn_credentials}
\alias{set_iucn_credentials}
\title{Store SpeciesLink credential}
\usage{
set_iucn_credentials(iucn_key, verbose = TRUE)
}
\arguments{
\item{iucn_key}{(character) your IUCN API key. See Details.}

\item{verbose}{(logical) if \code{TRUE}, prints messages about the progress and
the number of species being checked. Default is \code{TRUE}.}
}
\value{
No return value.
}
\description{
This function sets the IUCN API key as an environment variable for the
current R session. This key is required to obtain distributional data
from IUCN.
}
\details{
To check your API key, visit: \url{https://api.iucnredlist.org/users/edit}.
}
\examples{
\dontrun{
set_iucn_credentials(iucn_key = "my_key")
}

}
