% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_iucn_credentials.R
\name{set_iucn_credentials}
\alias{set_iucn_credentials}
\title{Store SpeciesLink credential}
\usage{
set_iucn_credentials(iucn_key, overwrite = FALSE,
                            open_Renviron = FALSE)
}
\arguments{
\item{iucn_key}{(character) your SpeciesLink API key. See Details.}

\item{overwrite}{(logical) whether to overwrite IUCN credential if it
already exists. Default is FALSE.}

\item{open_Renviron}{(logical) whether to open the .Renviron file after
saving the credentials. Default is FALSE.}
}
\value{
If \code{open_Renviron} is set to TRUE, it opens the .Renviron file. Otherwise,
the credentials are saved silently.
}
\description{
This function stores the IUCN credential (API key) in the R environment. This
API key is required to obtain distributional data from IUCN.
}
\details{
To check your API key, visit: \url{https://api.iucnredlist.org/users/edit}.
}
\examples{
\dontrun{
set_iucn_credentials(iucn_key = "my_key")
}

}
