% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_specieslink_credentials.R
\name{set_specieslink_credentials}
\alias{set_specieslink_credentials}
\title{Store SpeciesLink credential}
\usage{
set_specieslink_credentials(specieslink_key, verbose = TRUE)
}
\arguments{
\item{specieslink_key}{(character) your SpeciesLink API key.}

\item{verbose}{(logical) if \code{TRUE}, prints messages about the progress and
the number of species being checked. Default is \code{TRUE}.}
}
\value{
No return value.
}
\description{
This function sets the SpeciesLink API key as an environment variable for
the current R session. This API key is required to retrieve occurrence records from
SpeciesLink.
}
\details{
To check your API key, visit: \url{https://specieslink.net/aut/profile/apikeys}.
}
\examples{
\dontrun{
set_specieslink_credentials(specieslink_key = "my_key")
}

}
