% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialize.R
\name{spatialize}
\alias{spatialize}
\title{Title}
\usage{
spatialize(
  occ,
  long = "decimalLongitude",
  lat = "decimalLatitude",
  crs = "epsg:4326",
  force_numeric = TRUE
)
}
\arguments{
\item{force_numeric}{}
}
\description{
Title
}
